requirements.txt
fastapi
uvicorn
pydantic[email]
pymongo
motor
python-decouple
python-multipart
boto3
httpx
pathlib
